Changed index.js to app.js moved database.js
